+++ src/common/OpenEdgeJsonReaded.ts
@@ -2,25 +2,23 @@ import { PfParser } from '../view/app/utils/PfParser';
 import { IConfig } from '../view/app/model';
 import * as path from 'path';
 import * as fs from 'fs';
+import stripjsoncomments from 'strip-json-comments';
+
 
 export function readFile(fileName: string): string {
-    while (fileName.charAt(0) === '/') {
-        fileName = fileName.substring(1);
-    }
-
     const allFileContents = fs.readFileSync(fileName, 'utf-8');
     return allFileContents;
 }
 
-export function getOEVersion(fileContent: string){
-    const data = JSON.parse(fileContent);
-    const  {oeversion} = data;
-    
+export function getOEVersion(fileContent: string) {
+    const data = JSON.parse(stripjsoncomments(fileContent));
+    const { oeversion } = data;
+
     return oeversion;
 }
 
 export function parseOEFile(fileContent: string, filePath: string) {
-    const data = JSON.parse(fileContent);
+    const data = JSON.parse(stripjsoncomments(fileContent));
     const { name, dbConnections } = data;
     const configList: IConfig[] = [];